Move icon to icons folder
