+++ src/lib/icons/filter_icon.svelte
